pyproject.toml
[tool.poetry]
name = "ai-programming"
version = "0.1.0"
description = ""
authors = ["Mathias Aas <mathias.aas123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.x"
pycairo = "^1.26.0"
pillow = "^10.3.0"
tensorflow = "2.13"
tensorflow-macos = "2.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
